* use Rails::Railtie * added migrate option
