--- init.rb
@@ -1,1 +1,0 @@
-require File.dirname(__FILE__) + "/rails/init"
