+++ lib/memory_test_fix/railtie.rb
@@ -1,0 +1,7 @@
+module MemoryTestFix
+  class Railtie < ::Rails::Railtie
+    config.before_eager_load do
+      MemoryTestFix::SchemaLoader.init_schema
+    end
+  end
+end
