--- rails/init.rb
@@ -1,2 +1,0 @@
-
-require 'memory_test_fix'
